Rename sleep util file and move debounce into it
