--- src/utils/sleep.ts
@@ -1,3 +1,0 @@
-export function sleep(timeout = 1000): Promise<void> {
-    return new Promise(r => setTimeout(r, timeout));
-}
